+++ example/poisson/poisson_driver.cpp
@@ -18,7 +18,7 @@
 
 // Local Includes
 #include "amr_criteria/refinement_package.hpp"
-#include "bvals/cc/bvals_cc_in_one.hpp"
+#include "bvals/comms/bvals_in_one.hpp"
 #include "interface/metadata.hpp"
 #include "interface/update.hpp"
 #include "mesh/meshblock_pack.hpp"
@@ -150,8 +150,7 @@ TaskCollection PoissonDriver::MakeTaskCollection(BlockList_t &blocks) {
     solver.SetFailWithMaxIterations(fail_flag);
     solver.SetWarnWithMaxIterations(warn_flag);
 
-    auto start_recv = solver.AddTask(
-        none, parthenon::cell_centered_bvars::StartReceiveBoundaryBuffers, md);
+    auto start_recv = solver.AddTask(none, parthenon::StartReceiveBoundaryBuffers, md);
 
     auto update = solver.AddTask(mat_elem, poisson_package::UpdatePhi<MeshData<Real>>,
                                  md.get(), mdelta.get());
@@ -178,12 +177,11 @@ TaskCollection PoissonDriver::MakeTaskCollection(BlockList_t &blocks) {
                                      &update_norm.val)
                                : none);
 
-    auto send = solver.AddTask(update, cell_centered_bvars::SendBoundaryBuffers, md);
+    auto send = solver.AddTask(update, SendBoundaryBuffers, md);
 
-    auto recv =
-        solver.AddTask(start_recv, cell_centered_bvars::ReceiveBoundaryBuffers, md);
+    auto recv = solver.AddTask(start_recv, ReceiveBoundaryBuffers, md);
 
-    auto setb = solver.AddTask(recv | update, cell_centered_bvars::SetBoundaries, md);
+    auto setb = solver.AddTask(recv | update, SetBoundaries, md);
 
     auto check = solver.SetCompletionTask(
         send | setb | report_norm, poisson_package::CheckConvergence<MeshData<Real>>,