+++ src/outputs/ascent.cpp
@@ -169,7 +169,7 @@ void AscentOutput::WriteOutputFile(Mesh *pm, ParameterInput *pin, SimTime *tm,
     // create a field for each component of each variable pack
     auto &mbd = pmb->meshblock_data.Get();
 
-    for (const auto &var : mbd->GetCellVariableVector()) {
+    for (const auto &var : mbd->GetVariableVector()) {
       // ensure that only cell vars are added (for now) as the topology above is only
       // valid for cell centered vars
       if (!var->IsSet(Metadata::Cell)) {