+++ src/prolong_restrict/pr_loops.hpp
@@ -24,7 +24,7 @@
 #include <utility> // std::forward
 #include <vector>
 
-#include "bvals/cc/bnd_info.hpp"       // for buffercache_t
+#include "bvals/comms/bnd_info.hpp"    // for buffercache_t
 #include "coordinates/coordinates.hpp" // for coordinates
 #include "globals.hpp"                 // for Globals
 #include "kokkos_abstraction.hpp"      // for ParArray
@@ -74,10 +74,9 @@ GetLoopBoundsFromBndInfo(const Info_t &info, const int ckbs, const int cjbs, int
 // the size of the buffer cache.
 template <int DIM, class Stencil>
 inline void
-ProlongationRestrictionLoop(const cell_centered_bvars::BufferCache_t &info,
-                            const Idx_t &buffer_idxs, const IndexShape &cellbounds,
-                            const IndexShape &c_cellbounds, const RefinementOp_t op,
-                            const std::size_t nbuffers) {
+ProlongationRestrictionLoop(const BufferCache_t &info, const Idx_t &buffer_idxs,
+                            const IndexShape &cellbounds, const IndexShape &c_cellbounds,
+                            const RefinementOp_t op, const std::size_t nbuffers) {
   const IndexDomain interior = IndexDomain::interior;
   auto ckb = c_cellbounds.GetBoundsK(interior);
   auto cjb = c_cellbounds.GetBoundsJ(interior);
@@ -109,7 +108,7 @@ ProlongationRestrictionLoop(const cell_centered_bvars::BufferCache_t &info,
 }
 template <int DIM, class Stencil>
 inline void
-ProlongationRestrictionLoop(const cell_centered_bvars::BufferCacheHost_t &info_h,
+ProlongationRestrictionLoop(const BufferCacheHost_t &info_h,
                             const IdxHost_t &buffer_idxs_h, const IndexShape &cellbounds,
                             const IndexShape &c_cellbounds, const RefinementOp_t op,
                             const std::size_t nbuffers) {
@@ -145,8 +144,7 @@ ProlongationRestrictionLoop(const cell_centered_bvars::BufferCacheHost_t &info_h
 }
 template <int DIM, class Stencil>
 inline void
-ProlongationRestrictionLoop(const cell_centered_bvars::BufferCache_t &info,
-                            const cell_centered_bvars::BufferCacheHost_t &info_h,
+ProlongationRestrictionLoop(const BufferCache_t &info, const BufferCacheHost_t &info_h,
                             const Idx_t &buffer_idxs, const IdxHost_t &buffer_idxs_h,
                             const IndexShape &cellbounds, const IndexShape &c_cellbounds,
                             const RefinementOp_t op, const std::size_t nbuffers) {